Revert "feat: add trade settlement and refactor code"
