+++ group-buy-trade-platform-app/src/main/resources/mybatis/mapper/group_buy_order_mapper.xml
@@ -37,34 +37,15 @@
         ]]>
     </update>
 
-    <update id="updateAddCompleteCount" parameterType="java.lang.String">
-        <![CDATA[
-        update group_buy_order
-        set complete_count = complete_count + 1, update_time= now()
-        where team_id = #{teamId} and complete_count < target_count
-        ]]>
-    </update>
-
     <update id="updateSubtractionLockCount" parameterType="java.lang.String">
         update group_buy_order
         set lock_count = lock_count - 1, update_time= now()
         where team_id = #{teamId} and lock_count > 0
     </update>
 
-    <update id="updateOrderStatus2COMPLETE" parameterType="java.lang.String">
-        update group_buy_order
-        set status = 1, update_time= now()
-        where team_id = #{teamId} and status = 0
-    </update>
-
     <select id="queryGroupBuyProgress" parameterType="java.lang.String" resultMap="dataMap">
         select target_count, complete_count, lock_count from group_buy_order
         where team_id = #{teamId}
-    </select>
-
-    <select id="queryGroupBuyTeamByTeamId" parameterType="java.lang.String" resultMap="dataMap">
-        select team_id, activity_id, target_count, complete_count, lock_count, status
-        from group_buy_order where team_id = #{teamId}
     </select>
 
 </mapper>
