--- group-buy-trade-platform-app/src/main/resources/mybatis/mapper/notify_task_mapper.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.tinuvile.infrastructure.dao.INotifyTaskDao">
-
-    <resultMap id="dataMap" type="com.tinuvile.infrastructure.dao.po.NotifyTask">
-        <id column="id" property="id"/>
-        <result column="activity_id" property="activityId"/>
-        <result column="team_id" property="teamId"/>
-        <result column="notify_url" property="notifyUrl"/>
-        <result column="notify_count" property="notifyCount"/>
-        <result column="notify_status" property="notifyStatus"/>
-        <result column="parameter_json" property="parameterJson"/>
-        <result column="create_time" property="createTime"/>
-        <result column="update_time" property="updateTime"/>
-    </resultMap>
-
-    <insert id="insert" parameterType="com.tinuvile.infrastructure.dao.po.NotifyTask">
-        insert into notify_task(activity_id, team_id, notify_url, notify_count, notify_status, parameter_json,
-                                create_time, update_time)
-        values(#{activityId}, #{teamId}, #{notifyUrl}, #{notifyCount}, #{notifyStatus}, #{parameterJson}, now(),now())
-    </insert>
-
-</mapper>
