--- group-buy-trade-platform-app/src/test/java/com/tinuvile/test/domain/trade/TradeSettlementOrderServiceTest.java
@@ -1,41 +1,0 @@
-package com.tinuvile.test.domain.trade;
-
-
-import com.alibaba.fastjson.JSON;
-import com.tinuvile.domain.trade.model.entity.TradePaySettlementEntity;
-import com.tinuvile.domain.trade.model.entity.TradePaySuccessEntity;
-import com.tinuvile.domain.trade.service.ITradeSettlementOrderService;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-import javax.annotation.Resource;
-
-/**
- * @author Tinuvile
- * @description
- * @since 2025/11/14
- */
-@Slf4j
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class TradeSettlementOrderServiceTest {
-
-    @Resource
-    private ITradeSettlementOrderService tradeSettlementOrderService;
-
-    @Test
-    public void test_settlementMarketPayOrder() {
-        TradePaySuccessEntity tradePaySuccessEntity = new TradePaySuccessEntity();
-        tradePaySuccessEntity.setSource("s01");
-        tradePaySuccessEntity.setChannel("c01");
-        tradePaySuccessEntity.setUserId("tin02");
-        tradePaySuccessEntity.setOutTradeNo("451517755304");
-        TradePaySettlementEntity tradePaySettlementEntity = tradeSettlementOrderService.settlementMarketPayOrder(tradePaySuccessEntity);
-        log.info("请求参数:{}", JSON.toJSONString(tradePaySuccessEntity));
-        log.info("响应参数:{}", JSON.toJSONString(tradePaySettlementEntity));
-    }
-
-}
