+++ group-buy-trade-platform-domain/src/main/java/com/tinuvile/domain/tag/service/TagService.java
@@ -37,8 +37,6 @@ public class TagService implements ITagService {
                 add("Erchamion");
                 add("TestUser001");
                 add("tin01");
-                add("tin02");
-                add("tin03");
             }
         };
