+++ group-buy-trade-platform-domain/src/main/java/com/tinuvile/domain/trade/adapter/repository/ITradeRepository.java
@@ -2,15 +2,13 @@ package com.tinuvile.domain.trade.adapter.repository;
 
 
 import com.tinuvile.domain.trade.model.aggregate.GroupBuyOrderAggregate;
-import com.tinuvile.domain.trade.model.aggregate.GroupBuyTeamSettlementAggregate;
 import com.tinuvile.domain.trade.model.entity.GroupBuyActivityEntity;
-import com.tinuvile.domain.trade.model.entity.GroupBuyTeamEntity;
 import com.tinuvile.domain.trade.model.entity.MarketPayOrderEntity;
 import com.tinuvile.domain.trade.model.valobj.GroupBuyProcessVO;
 
 /**
  * @author Tinuvile
- * @description 交易仓储服务接口
+ * @description
  * @since 2025/11/11
  */
 public interface ITradeRepository {
@@ -24,9 +22,5 @@ public interface ITradeRepository {
     GroupBuyActivityEntity queryGroupBuyActivityEntityByActivityId(Long activityId);
 
     Integer queryOrderCountByActivityId(Long activityId, String userId);
-
-    GroupBuyTeamEntity queryGroupBuyTeamByTeamId(String teamId);
-
-    void settlementMarketPayOrder(GroupBuyTeamSettlementAggregate groupBuyTeamSettlementAggregate);
 
 }
