--- group-buy-trade-platform-domain/src/main/java/com/tinuvile/domain/trade/model/entity/GroupBuyTeamEntity.java
@@ -1,39 +1,0 @@
-package com.tinuvile.domain.trade.model.entity;
-
-
-import com.tinuvile.types.enums.GroupBuyOrderEnumVO;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * @author Tinuvile
- * @description 拼团组队实体
- * @since 2025/11/14
- */
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class GroupBuyTeamEntity {
-
-    /** 拼单组队ID */
-    private String teamId;
-
-    /** 活动ID */
-    private Long activityId;
-
-    /** 目标数量 */
-    private Integer targetCount;
-
-    /** 完成数量 */
-    private Integer completeCount;
-
-    /** 锁单数量 */
-    private Integer lockCount;
-
-    /** 状态（0-拼单中、1-完成、2-失败） */
-    private GroupBuyOrderEnumVO status;
-
-}
