--- group-buy-trade-platform-domain/src/main/java/com/tinuvile/domain/trade/service/settlement/TradeSettlementOrderService.java
@@ -1,60 +1,0 @@
-package com.tinuvile.domain.trade.service.settlement;
-
-
-import com.tinuvile.domain.trade.adapter.repository.ITradeRepository;
-import com.tinuvile.domain.trade.model.aggregate.GroupBuyTeamSettlementAggregate;
-import com.tinuvile.domain.trade.model.entity.*;
-import com.tinuvile.domain.trade.service.ITradeSettlementOrderService;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-import javax.annotation.Resource;
-
-/**
- * @author Tinuvile
- * @description
- * @since 2025/11/14
- */
-@Slf4j
-@Service
-public class TradeSettlementOrderService implements ITradeSettlementOrderService {
-
-    @Resource
-    private ITradeRepository repository;
-
-    @Override
-    public TradePaySettlementEntity settlementMarketPayOrder(TradePaySuccessEntity tradePaySuccessEntity) {
-        log.info("拼团交易 - 支付订单结算:{} outTradeNo:{}", tradePaySuccessEntity.getUserId(), tradePaySuccessEntity.getOutTradeNo());
-
-        // 查询拼团信息
-        MarketPayOrderEntity marketPayOrderEntity = repository.queryMarketPayOrderEntityByOutTradeNo(tradePaySuccessEntity.getUserId(), tradePaySuccessEntity.getOutTradeNo());
-        if (null == marketPayOrderEntity) {
-            log.info("不存在的外部交易单号或用户已退单，不需要做支付订单结算:{} outTradeNo:{}", tradePaySuccessEntity.getUserId(), tradePaySuccessEntity.getOutTradeNo());
-            return null;
-        }
-
-        // 查询组团信息
-        GroupBuyTeamEntity groupBuyTeamEntity = repository.queryGroupBuyTeamByTeamId(marketPayOrderEntity.getTeamId());
-
-        // 构建聚合对象
-        GroupBuyTeamSettlementAggregate groupBuyTeamSettlementAggregate = GroupBuyTeamSettlementAggregate.builder()
-                .userEntity(UserEntity.builder().userId(tradePaySuccessEntity.getUserId()).build())
-                .groupBuyTeamEntity(groupBuyTeamEntity)
-                .tradePaySuccessEntity(tradePaySuccessEntity)
-                .build();
-
-        // 拼团交易结算
-        repository.settlementMarketPayOrder(groupBuyTeamSettlementAggregate);
-
-        // 返回结算信息
-        return TradePaySettlementEntity.builder()
-                .source(tradePaySuccessEntity.getSource())
-                .channel(tradePaySuccessEntity.getChannel())
-                .userId(tradePaySuccessEntity.getUserId())
-                .teamId(marketPayOrderEntity.getTeamId())
-                .activityId(groupBuyTeamEntity.getActivityId())
-                .outTradeNo(tradePaySuccessEntity.getOutTradeNo())
-                .build();
-    }
-
-}
