+++ group-buy-trade-platform-infrastructure/src/main/java/com/tinuvile/infrastructure/adapter/repository/TradeRepository.java
@@ -1,23 +1,18 @@
 package com.tinuvile.infrastructure.adapter.repository;
 
 
-import com.alibaba.fastjson.JSON;
 import com.tinuvile.domain.trade.adapter.repository.ITradeRepository;
 import com.tinuvile.domain.trade.model.aggregate.GroupBuyOrderAggregate;
-import com.tinuvile.domain.trade.model.aggregate.GroupBuyTeamSettlementAggregate;
 import com.tinuvile.domain.trade.model.entity.*;
 import com.tinuvile.domain.trade.model.valobj.GroupBuyProcessVO;
 import com.tinuvile.domain.trade.model.valobj.TradeOrderStatusEnumVO;
 import com.tinuvile.infrastructure.dao.IGroupBuyActivityDao;
 import com.tinuvile.infrastructure.dao.IGroupBuyOrderDao;
 import com.tinuvile.infrastructure.dao.IGroupBuyOrderListDao;
-import com.tinuvile.infrastructure.dao.INotifyTaskDao;
 import com.tinuvile.infrastructure.dao.po.GroupBuyActivity;
 import com.tinuvile.infrastructure.dao.po.GroupBuyOrder;
 import com.tinuvile.infrastructure.dao.po.GroupBuyOrderList;
-import com.tinuvile.infrastructure.dao.po.NotifyTask;
 import com.tinuvile.types.common.Constants;
-import com.tinuvile.types.enums.GroupBuyOrderEnumVO;
 import com.tinuvile.types.enums.ResponseCode;
 import com.tinuvile.types.exception.AppException;
 import lombok.extern.slf4j.Slf4j;
@@ -28,8 +23,6 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
-import java.util.HashMap;
-import java.util.List;
 
 /**
  * @author Tinuvile
@@ -49,9 +42,6 @@ public class TradeRepository implements ITradeRepository {
     @Resource
     private IGroupBuyOrderListDao groupBuyOrderListDao;
 
-    @Resource
-    private INotifyTaskDao notifyTaskDao;
-
     @Override
     public MarketPayOrderEntity queryMarketPayOrderEntityByOutTradeNo(String userId, String outTradeNo) {
         GroupBuyOrderList groupBuyOrderListReq = new GroupBuyOrderList();
@@ -64,7 +54,6 @@ public class TradeRepository implements ITradeRepository {
         }
 
         return MarketPayOrderEntity.builder()
-                .teamId(groupBuyOrderListRes.getTeamId())
                 .orderId(groupBuyOrderListRes.getOrderId())
                 .deductionPrice(groupBuyOrderListRes.getDeductionPrice())
                 .tradeOrderStatusEnumVO(TradeOrderStatusEnumVO.valueOf(groupBuyOrderListRes.getStatus()))
@@ -131,7 +120,6 @@ public class TradeRepository implements ITradeRepository {
 
         return MarketPayOrderEntity.builder()
                 .orderId(orderId)
-                .teamId(teamId)
                 .deductionPrice(payDiscountEntity.getDeductionPrice())
                 .tradeOrderStatusEnumVO(TradeOrderStatusEnumVO.CREATE)
                 .build();
@@ -178,72 +166,6 @@ public class TradeRepository implements ITradeRepository {
         groupBuyOrderListReq.setActivityId(activityId);
         groupBuyOrderListReq.setUserId(userId);
         return groupBuyOrderListDao.queryOrderCountByActivityId(groupBuyOrderListReq);
-
-    }
-
-    @Override
-    public GroupBuyTeamEntity queryGroupBuyTeamByTeamId(String teamId) {
-        GroupBuyOrder groupBuyOrder = groupBuyOrderDao.queryGroupBuyTeamByTeamId(teamId);
-        return GroupBuyTeamEntity.builder()
-                .teamId(groupBuyOrder.getTeamId())
-                .activityId(groupBuyOrder.getActivityId())
-                .targetCount(groupBuyOrder.getTargetCount())
-                .completeCount(groupBuyOrder.getCompleteCount())
-                .lockCount(groupBuyOrder.getLockCount())
-                .status(GroupBuyOrderEnumVO.valueOf(groupBuyOrder.getStatus()))
-                .build();
-    }
-
-    @Transactional(timeout = 500)
-    @Override
-    public void settlementMarketPayOrder(GroupBuyTeamSettlementAggregate groupBuyTeamSettlementAggregate) {
-
-        UserEntity userEntity = groupBuyTeamSettlementAggregate.getUserEntity();
-        GroupBuyTeamEntity groupBuyTeamEntity = groupBuyTeamSettlementAggregate.getGroupBuyTeamEntity();
-        TradePaySuccessEntity tradePaySuccessEntity = groupBuyTeamSettlementAggregate.getTradePaySuccessEntity();
-
-        // 更新拼团订单明细状态
-        GroupBuyOrderList groupBuyOrderListReq = new GroupBuyOrderList();
-        groupBuyOrderListReq.setUserId(userEntity.getUserId());
-        groupBuyOrderListReq.setOutTradeNo(tradePaySuccessEntity.getOutTradeNo());
-        int updateOrderListStatusCount = groupBuyOrderListDao.updateOrderStatus2COMPLETE(groupBuyOrderListReq);
-        if (1 != updateOrderListStatusCount) {
-            throw new AppException(ResponseCode.UPDATE_ZERO);
-        }
-
-        // 更新拼团达成数量
-        int updateAddCount = groupBuyOrderDao.updateAddCompleteCount(groupBuyTeamEntity.getTeamId());
-        if (1 != updateAddCount) {
-            throw new AppException(ResponseCode.UPDATE_ZERO);
-        }
-
-        // 更新拼团完成状态
-        if (groupBuyTeamEntity.getTargetCount() - groupBuyTeamEntity.getCompleteCount() == 1) {
-            int updateCompleteCount = groupBuyOrderDao.updateOrderStatus2COMPLETE(groupBuyTeamEntity.getTeamId());
-            if (1 != updateCompleteCount) {
-                throw new AppException(ResponseCode.UPDATE_ZERO);
-            }
-
-            // 查询拼团交易完成外部单号列表
-            List<String> outTradeNoList = groupBuyOrderListDao.queryGroupBuyCompleteOrderOutTradeNoListByTeamId(groupBuyTeamEntity.getTeamId());
-
-            // 拼团完成写入回调任务记录
-            NotifyTask notifyTask = new NotifyTask();
-            notifyTask.setActivityId(groupBuyTeamEntity.getActivityId());
-            notifyTask.setTeamId(groupBuyTeamEntity.getTeamId());
-            notifyTask.setNotifyUrl("暂无");
-            notifyTask.setNotifyCount(0);
-            notifyTask.setNotifyStatus(0);
-            notifyTask.setParameterJson(JSON.toJSONString(new HashMap<String, Object>() {
-                {
-                    put("teamId", groupBuyTeamEntity.getTeamId());
-                    put("outTradeNoList", outTradeNoList);
-                }
-            }));
-
-            notifyTaskDao.insert(notifyTask);
-
-        }
 
     }
