+++ group-buy-trade-platform-infrastructure/src/main/java/com/tinuvile/infrastructure/dao/IGroupBuyOrderListDao.java
@@ -4,11 +4,9 @@ package com.tinuvile.infrastructure.dao;
 import com.tinuvile.infrastructure.dao.po.GroupBuyOrderList;
 import org.apache.ibatis.annotations.Mapper;
 
-import java.util.List;
-
 /**
  * @author Tinuvile
- * @description 用户拼单明细
+ * @description
  * @since 2025/11/11
  */
 @Mapper
@@ -19,9 +17,5 @@ public interface IGroupBuyOrderListDao {
     GroupBuyOrderList queryGroupBuyOrderRecordByOutTradeNo(GroupBuyOrderList groupBuyOrderListReq);
 
     Integer queryOrderCountByActivityId(GroupBuyOrderList groupBuyOrderListReq);
-
-    int updateOrderStatus2COMPLETE(GroupBuyOrderList groupBuyOrderListReq);
-
-    List<String> queryGroupBuyCompleteOrderOutTradeNoListByTeamId(String teamId);
 
 }
