--- group-buy-trade-platform-infrastructure/src/main/java/com/tinuvile/infrastructure/dao/INotifyTaskDao.java
@@ -1,17 +1,0 @@
-package com.tinuvile.infrastructure.dao;
-
-
-import com.tinuvile.infrastructure.dao.po.NotifyTask;
-import org.apache.ibatis.annotations.Mapper;
-
-/**
- * @author Tinuvile
- * @description 回调任务
- * @since 2025/11/14
- */
-@Mapper
-public interface INotifyTaskDao {
-
-    void insert(NotifyTask notifyTask);
-
-}
