+++ group-buy-trade-platform-types/src/main/java/com/tinuvile/types/enums/ResponseCode.java
@@ -13,7 +13,6 @@ public enum ResponseCode {
     UN_ERROR("0001", "未知失败"),
     ILLEGAL_PARAMETER("0002", "非法参数"),
     INDEX_DUPLICATE("0003", "唯一索引重复"),
-    UPDATE_ZERO("0004", "更新记录为0"),
 
     E0001("E0001", "不存在对应的折扣计算服务"),
     E0002("E0002", "无拼团营销配置"),
